kevin/sunday-stuff (#10)
* feat(action): improve barcode scanning and state management

Refactor the barcode scanning functionality to use the `isFocused` hook to
ensure the `onBarcodeScanned` callback is only called when the screen is
focused.

Additionally, reorganize the placement of the `DisplayJSON` component and
the "Reset local state" button within the "Dev Toolbox" section.

* feat: width of wallet transaction date

Modify the wallet transaction date text to ensure it fits within the
available width. This is achieved by setting the `numberOfLines` prop to
1 and enabling `adjustsFontSizeToFit` on the `Text` component.

* feat: disable dev toolbox by default

1. Disabled the `enableDevToolbox` feature by default; removed the commented-out code that previously enabled the dev toolbox based on the `NODE_ENV` environment variable.
2. Simplified the `uriValidator` function by removing the unnecessary `console.log` statement.